Update activity tests to fragment tests
